+++ src/main/java/nextstep/step2/LadderGame.java
@@ -1,5 +1,10 @@
 package nextstep.step2;
 
+import nextstep.step2.domain.Ladder;
+import nextstep.step2.domain.People;
+import nextstep.step2.view.InputView;
+import nextstep.step2.view.OutputView;
+
 public class LadderGame {
 
     public static void main(String[] args) {
