+++ src/main/java/nextstep/step2/domain/Line.java
@@ -1,4 +1,4 @@
-package nextstep.step2;
+package nextstep.step2.domain;
 
 import java.util.Objects;
 import java.util.Random;
