+++ src/main/java/nextstep/step2/domain/People.java
@@ -1,4 +1,4 @@
-package nextstep.step2;
+package nextstep.step2.domain;
 
 import java.util.Arrays;
 import java.util.List;
