+++ src/main/java/nextstep/step2/exception/Error.java
@@ -1,4 +1,4 @@
-package nextstep.step2;
+package nextstep.step2.exception;
 
 public class Error {
