+++ src/test/java/nextstep/step2/LinesTest.java
@@ -1,11 +1,12 @@
 package nextstep.step2;
 
+import nextstep.step2.domain.Line;
+import nextstep.step2.domain.Lines;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
-import static java.lang.Boolean.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
