+++ src/test/java/nextstep/step2/PeopleTest.java
@@ -1,7 +1,8 @@
 package nextstep.step2;
 
+import nextstep.step2.domain.People;
+import nextstep.step2.domain.Person;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
